refactor: moved feedback widget script into a seperate js file
