+++ src/learning-header/LearningHeader.jsx
@@ -6,6 +6,7 @@ import {
 } from '@edx/frontend-platform';
 import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
 import { AppContext, AppProvider } from '@edx/frontend-platform/react';
+import classNames from 'classnames';
 import AnonymousUserMenu from './AnonymousUserMenu';
 import AuthenticatedUserDropdown from './AuthenticatedUserDropdown';
 import messages from './messages';
@@ -76,7 +77,12 @@ const LearningHeader = ({
         <a className="sr-only sr-only-focusable" href="#main-content">{intl.formatMessage(messages.skipNavLink)}</a>
         <div className="px-4 py-2.5 d-flex align-items-center learning-header-container">
           {headerLogo}
-          <div className="flex-grow-1 course-title-lockup" style={{ lineHeight: 1 }}>
+          <div
+            className={classNames('flex-grow-1', {
+              'course-title-lockup': courseTitle || courseOrg || courseNumber,
+            })}
+            style={{ lineHeight: 1 }}
+          >
             <span className="d-block small m-0">{courseOrg} {courseNumber}</span>
             <span className="d-block m-0 font-weight-bold course-title">{courseTitle}</span>
           </div>
